trading-bot-arena/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
supabase==2.4.6
ccxt==4.3.6
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-json-logger==2.0.7
